+++ MonoGameTest.Common/Packets/MoveCharacterPacket.cs
@@ -2,8 +2,8 @@ namespace MonoGameTest.Common {
 
 	public class MoveCharacterPacket : ICharacterPacket {
 		public int OriginCharacterId { get; set; }
-		public int X { get; set; }
-		public int Y { get; set; }
+		public long X { get; set; }
+		public long Y { get; set; }
 		public float Duration { get; set; }
 	}
