+++ MonoGameTest.Common/Packets/MoveCommand.cs
@@ -1,8 +1,8 @@
 namespace MonoGameTest.Common {
 
 	public class MoveCommand {
-		public int X { get; set; }
-		public int Y { get; set; }
+		public long X { get; set; }
+		public long Y { get; set; }
 	}
 
 }
