+++ MonoGameTest.Common/Packets/ProjectilePatcket.cs
@@ -1,8 +1,8 @@
 namespace MonoGameTest.Common {
 
 	public class ProjectilePacket {
-		public int OriginX { get; set; }
-		public int OriginY { get; set; }
+		public long OriginX { get; set; }
+		public long OriginY { get; set; }
 		public int TargetCharacterId { get; set; }
 		public int SkillId { get; set; }
 	}
