+++ MonoGameTest.Common/Packets/SkillStartFixedPacket.cs
@@ -3,8 +3,8 @@ namespace MonoGameTest.Common {
 	public class SkillStartFixedPacket {
 		public int SkillId { get; set; }
 		public int OriginCharacterId { get; set; }
-		public int TargetX { get; set; }
-		public int TargetY { get; set; }
+		public long TargetX { get; set; }
+		public long TargetY { get; set; }
 	}
 
 }
