+++ MonoGameTest.Common/Packets/TargetFixedPacket.cs
@@ -2,8 +2,8 @@ namespace MonoGameTest.Common {
 
 	public class TargetFixedPacket {
 		public int CharacterId { get; set; }
-		public int X { get; set; }
-		public int Y { get; set; }
+		public long X { get; set; }
+		public long Y { get; set; }
 	}
 
 }
